merge

+++ Segment.h
@@ -3,12 +3,12 @@ using Eigen::Vector3d;
 
 class Segment {
   public:
-    Segment(int length) : length(length) {
+    Segment(float length) : length(length) {
         rot = Vector3d(0,0,0);
         numSegments += 1;
     }
     Vector3d rot;
-    int length;
+    float length;
     static int numSegments;
     static std::vector<Segment> * segments;
 };
